requirements.txt
requests_oauthlib
websocket-client
websockets
hydra-core
google-cloud-bigquery
google-cloud-logging
google-cloud-storage
rel
evaluate
opentelemetry-api
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation
opentelemetry-instrumentation-requests
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
pydash
lxml
